removed int16 from hist (from utils) and refactored pcap_classification to work with the neew handeling of fllow and block

+++ misc/utils.py
@@ -20,7 +20,7 @@ def build_pic(stream: Sequence[Tuple[float, int]],
     # scaling stream_duration to pic's width
     x_axis_to_second_ratio = pic_width * 1.0 / stream_duration_in_seconds
 
-    hist = torch.zeros(pic_width, pic_height, dtype=torch.int16)
+    hist = torch.zeros(pic_width, pic_height)
     for packet in stream:
         # packet is (time, size)
         x_position = int(floor(float(packet[0]) * x_axis_to_second_ratio))
